requirements.txt
opencv-python>=4.5.0
PyQt5>=5.15
moviepy>=1.0.3
librosa>=0.9.2
numpy>=1.21
scipy>=1.7
soundfile>=0.10
pydub>=0.25
ffmpeg-python>=0.2.0
pillow>=9.0
simpleaudio>=1.0.0
sounddevice>=0.4.6
pytest>=6.0
